fix(rds): Make most DatabaseClusterAttributes properties optional (#10291)
The `IDatabaseCluster` interface has a lot of properties, putting a large strain
on users who want to import clusters via `DatabaseClusterAttributes`. This
alters the attributes interface to contain almost exclusively optional props,
and lazily throw if any missing properties are accessed.

I believe this is a backwards-compatible change, but someone please double-check
me on that.

fixes #3587


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*