feat(rds): change excludeCharacters for Cluster, Instance and DatabaseSecret
Change the default excludeCharacters for Cluster,
Instance and DatabaseSecret to the character set " %+~`#$&*()|[]{}:;<>?!'/@\"\\",
as the previous set ('"@/\\')
had a tendency to generate problematic passwords that wouldn't work in the shell,
or with services like DMS.
Do the same for single- and multi-user rotations in Cluster and Instance as well.
Also allow passing a custom excludeCharacters for Credentials and SnapshotCredentials,
and also in addSingleUserRotation and addMultiUserRotation.

Fixes #4144

BREAKING CHANGE: the default excludeCharacters set for Instance, Cluster and DatabaseSecret is now " %+~`#$&*()|[]{}:;<>?!'/@\"\\"
* **rds**: the default excludeCharacters for addSingleUserRotation and addMultiUserRotation is now " %+~`#$&*()|[]{}:;<>?!'/@\"\\"
* **rds**: Instance.addSingleUserRotation now takes options as the first argument, instead of just Duration
* **rds**: Cluster.addSingleUserRotation now takes options as the first argument, instead of just Duration