fix(pipelines): make CdkPipeline build stage optional
CdkPipeline works perfectly fine without build stage, if already-built
cloud assembly is provided in the source stage (e.g. S3 source action).
A use case for this is separate CI and CD logic, where CDK
synthetization happens within the CI build and the assembly is stored as
an artefact to be deployed by a pipeline. This commit makes the build
stage optional, so that this use case is possible without a "dummy build
stage" to satisfy validation.