Let user pass EventRole to ScheduledFargateTask
aws-ecs-patterns/scheduled-fargate-task creates a new aws-events-targets/ecs-task to attach the task definition to a schedule. The EcsTask object accepts an optional `role` prop; defaulting to a newly created singleton role if not passed. This change allows the user to optionally specify an `eventRole` at the construction of `ScheduledFargateTask` object to pass through to the `addTaskDefinitionToEventTarget` method