chore: add yarn upgrade workflow (#10346)
This workflow runs `npm-check-updates` throughout the repository, and runs
an equivalent of `yarn install --force` (equivalent to `yarn install && yarn upgrade`,
except in a single command so more efficient).

Some packages enjoy "special handling":
- `@types/node` will stay on the currently set *major* version (since this reflects
  the minimal supported `node` version)
- `@types/fs-extra` will stay on the currently set *major* version (because the
  latest major version is known to be broken with `@types/node >=10`, as seen
  in DefinitelyTyped/DefinitelyTyped#30398
- `typescript` will stay on the currently set *minor* version (since it does not
  follow semantic versioning, and *Major.Minor* is where the stability point is

This ensures all dependencies are up-to-date, on the latest version, including
transitive dependencies that may not have been updated by the `dependabot`
workflows previously.

This files a pull request once done if there are any changes. It requires a secret
called `AUTOMATION_GITHUB_TOKEN` that will need to be created, and contain
a GitHub token that has `workflows` permissions, so PR validation workflows will
trigger on the PR.