fix(cli): `--profile` is ignored if AWS_ variables are set (#10362)
Change behavior to match the behavior of the AWS CLI. If `--profile` is
set, the given profile from the INI files will be used to the exclusion
of everything else.

If `--profile` is not set, the regular ambient credential process
is followed.

`AWS_PROFILE` is part of ambient credentials, and setting it does not
lead to triggering the new behavior. This is consistent with how the AWS
CLI behaves.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*