fix: SSM Association 'parameters' property has incorrect type (#10316)
There is a problem with the `Parameters` property of the `AWS::SSM::Association` resource type.
Server-side, its type is `Map<String, List<String>>`.
Unfortunately, that type is impossible to represent in the current CloudFormation specification that our code generation is based on.

Add a patch that changes the `Parameters` property of `AWS::SSM::Association` to:

```json
    "AWS::SSM::Association": {
      "Properties": {
        "Parameters": {
          "Type": "Map",
          "ItemType": "List",
          "PrimitiveItemItemType": "String",
          "UpdateType": "Mutable",
          "Required": false
        }
      }
    }
```

That is not a legal CFN spec type.
But we add a new property type, `MapOfListsOfPrimitives`,
to the `cfnspec` module, that represents it.
We then modify our L1 code generation in `cfn2ts` to use the new type,
allowing `CfnAssociationProps.parameters` to have type `{ [key: string]: string[] }`.

Fixes #3092

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*