feat(ec2): user-defined subnet selectors (#10112)
Continuation of https://github.com/aws/aws-cdk/pull/8526

This change adds a feature to SubnetSelection which lets the user provide objects that implements the new `ISubnetSelector` interface. These objects will be used by the VPC class to choose the subnets from the VPC.

This commit also provides an implementation of an ISubnetSelector that lets the user select subnets that contain IP addresses.

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*