fix(diff): fix the issue (#10322)
Fix the issue (10322) for `deepEqual`.

The cause of this issue was that checking the diffs in the `DependsOn` property returned a result of` dependsOnEqual`.

If the result of `dependsOnEqual` is false, there is no problem.
If the result of `dependsOnEqual` is true, need to check the diffs of other properties.