fix(codebuild): `encryptionKey` cannot be disabled
Because of a limitation of the CodeBuild Service API and its
CloudFormation implementation, it is not possible to disable the
encryption key used to encrypt uploaded artifacts after having deployed
a Project once that uses a key (the update back to "no key" is ignored).

Work around this by explicitly always selecting the default
`alias/aws/s3` key. This is the same one CodeBuild would have used if
no key was given, except it doesn't suffer from the "property cannot
be made empty" problem.