fix(core): CfnParameter of Number type cannot be used as a string (#10422)
CloudFormation allows for parameters of type 'Number' to be referenced,
using the 'Ref' keyword, into properties that are of type 'String'.

This will let customers now use CloudFormation maximum and minimum
constraints on the number parameter type, and still use the resulting
value in a property of string type.

fixes #10228


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*