fix(cli): stack outputs aren't sorted (#10328)
When running `cdk deploy` the stack outputs to the terminal are currently returned in the same order as the `describe stacks` API call, which does not seem to provide a contract on ordering, per the [docs](https://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_Stack.html).

This change sorts the keys of the stack outputs before display, which is consistent with "outputs" tab in the AWS CloudFormation console.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*