fix(pipelines): make CdkPipeline build stage optional (#10345)
In PR  #10148, @rix0rrr made it possible to provide a custom CodePipeline pipeline instance to CdkPipeline. This also made the `sourceAction` (Source stage) and `synthAction` (Build stage) props optional. 

However, validation was added to ensure that if `synthAction` is not provided, the pipeline already contains at least two stages (assuming that would be Source and Build).

Logically though, CdkPipeline works perfectly fine without Build stage, if an already-built cloud assembly is provided in the source stage (e.g. S3 source action). A use case for this is, for example, separating CI and CD logic, where CDK synthesis happens within the CI build and the assembly is stored as an artefact to be deployed by a pipeline. 

This PR makes the Build stage optional, to allow this use case without a need for a dummy build stage.

Example pipeline code:

```ts
export class PipelineStack extends cdk.Stack {
  constructor(scope: cdk.Construct, id: string, props: cdk.StackProps) {
    super(scope, id, props);

    const versionsBucket = new s3.Bucket(this, 'VersionsBucket', {
      bucketName: 's3pipeline-app-versions',
      versioned: true,
    });

    // The CodePipeline
    const cloudAssemblyArtifact = new codepipeline.Artifact()
    const codePipeline = new codepipeline.Pipeline(this, 'CodePipeline', {
      pipelineName: 'S3Pipeline',
      restartExecutionOnUpdate: true,
      stages: [{
        stageName: 'Source',
        actions: [new actions.S3SourceAction({
            actionName: 'S3',
            bucket: versionsBucket,
            bucketKey: 'cloudassembly.zip',
            output: cloudAssemblyArtifact
          })]
      }]
    });

    // CDK Pipeline
    const cdkPipeline = new pipelines.CdkPipeline(this, 'CdkPipeline', {
      codePipeline,
      cloudAssemblyArtifact,
    });

    // Add application stage
    cdkPipeline.addApplicationStage(new MyAppStage(this, "PreProd"));
  }
}
```


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*