feat(rds): support setting database master users from existing secrets
See https://github.com/aws/aws-cdk/issues/7927#issuecomment-694857345 for
motivation and design.

The current way of specifying master user logins for `DatabaseInstance` and
`DatabaseCluster` is inconsistent between the two and introduces some awkward
usage when creating a login from an existing `Secret`.

This change converts the existing `Login` interface (used by the `DatabaseCluster`)
into a class with factory methods for username/password or secret-based logins.
This also then re-uses that same interface for `DatabaseInstance`.

The one exception now will be `DatabaseInstanceFromSnapshot`, which has specific
requirements that deserved its own interface (`SnapshotLogin`).

As a side effect of this approach, existing `DatabaseCluster` users -- in
Typescript at least -- will not be broken. For example, the following are
equivalent:

```ts
new rds.DatabaseCluster(this, 'Cluster1', {
  // Existing usage
  masterUser: {
    username: 'admin',
  },
  // New usage
  masterUser: Login.fromUsername('admin'),
});
```

fixes #7927

BREAKING CHANGE: `DatabaseInstanceProps` and `DatabaseInstanceFromSnapshotProps` -
`masterUsername`, `masterUserPassword` and `masterUserPasswordEncryptionKey` moved
to `masterUser` as a new `Login` class.
* **rds:** `Login` converted from interface to class with factory methods. Use `Login.fromUsername` to replace
  existing usage.