fix(rds): allow creating Proxies for imported resources
The current ProxyTarget relied on the underlying L1s to get the engine type
for a given Cluster/Instance.
Change IDatabaseCluster and IInstanceEngine to add an (optional)
`engine` property that is used instead.
Allow the user to specify the engine when importing a Cluster or Instance.

Also move the logic of determining the engine family into IEngine.

Fixes #9195