chore(awslint): rules permit constructs to extend from 'constructs' module (#10472)
Introduce an environment variable - `AWSLINT_BASE_CONSTRUCT`
recognized by `awslint`. This environment variable indicates that the
module has [migrated][compat-rfc] away from construct classes and
interfaces from `@aws-cdk/core` module to those in `constructs`
module.

Specific rules in the linter recognize this variable and modify their
expectations.

Motivation
The primary motivation is to move the code base towards [removal of the
construct compat layer][compat-rfc] as part of [CDKv2].

A large number of code changes to adopt "constructs" module can already
be done as part of CDKv1 without incurring breaking changes to the API.

This change enables these changes to be performed module-by-module. As
modules are migrated, this flag will be enabled, to ensure no
regression.

[CDKv2]: https://github.com/aws/aws-cdk-rfcs/blob/master/text/0079-cdk-2.0.md
[compat-rfc]: https://github.com/aws/aws-cdk-rfcs/blob/master/text/0192-remove-constructs-compat.md

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*