fix(rds): allow creating Proxies for imported resources (#10488)
The current ProxyTarget relied on the underlying L1s to get the engine type
for a given Cluster/Instance.
Change IDatabaseCluster and IInstanceEngine to add an (optional)
`engine` property that is used instead.
Allow the user to specify the engine when importing a Cluster or Instance.

Also move the logic of determining the engine family into `IEngine`.

Fixes #9195

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*