chore(cli): undo rename 'versionReporting' => 'analyticsReporting' (#10587)
The switch `versionReporting` was renamed to `analyticsReporting`
in the framework in order to be more generic.

However, this rename was carried through in parts of the CLI as well,
especially in parts of the settings handling. This part is however
not strongly typed but *stringly* typed (remember that settings
are also read from `cdk.json`) and it's rather tricky to change there.

Undo this change as it's breaking the integration tests; for now,
the switch will still be called `versionReporting` in the CLI.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*