feat(rds): add the ability to exclude characters when generating passwords for Cluster, Instance, DatabaseSecret
Change the default excludeCharacters for Cluster,
Instance and DatabaseSecret to the character set ``" %+~`#$&*()|[]{}:;<>?!'/@\"\\"``,
as the previous set (`'"@/\\'`)
had a tendency to generate problematic passwords that wouldn't work in the shell,
or with services like DMS.
Do the same for single- and multi-user rotations in Cluster and Instance as well.
Also allow passing a custom excludeCharacters for Credentials and SnapshotCredentials,
and also in addSingleUserRotation and addMultiUserRotation.

Fixes #4144

BREAKING CHANGE: the default generated password exclude characters set for Instance, Cluster and `DatabaseSecret` is now ``" %+~`#$&*()|[]{}:;<>?!'/@\"\\"``
* **rds**: the default generated password exclude characters for `addSingleUserRotation()` and `addMultiUserRotation()` in Cluster and Instance is now ``" %+~`#$&*()|[]{}:;<>?!'/@\"\\"``
* **rds**: `Instance.addSingleUserRotation()` now takes options object as the first argument, instead of just `Duration`
* **rds**: `Cluster.addSingleUserRotation()` now takes options object as the first argument, instead of just `Duration`
* **rds**: `SnapshotCredentials.fromGeneratedPassword()` now takes an option object as the second argument, instead of just `IKey`

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*