feat(rds): the unversioned Cluster engine constants are no longer deprecated
The static constants in the DatabaseClusterEngine class were all deprecated in favor of using versioned engines.
However, it turns out that it's very common to use the unversioned engines for serverless Clusters,
for which the instances are fully managed, so updates are a lot less risky.
Because of that, un-deprecate the constants, instead adding a note that we don't recommend using them for non-serverless Clusters.