feat(elasticsearch): L2 for ElasticsearchDomain (#8369)
I was using cdk to stand up an Elasticsearch domain and thought I might as well generalise the effort.

Can I please get some feedback on the current implementation and if it is something you are happy to accept?

This implementation is still missing tests, more documentation and features like importing the resource. I will add them once I know I'm on the right track.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*