fix(core,cx-api): docker bundling fails during tests on macOS (#10620)
When running unit tests `outdir` defaults to a temporary directory in
the system temp directory. On macOS `os.tmpdir()` is a symlink. In 
`v1.61.0` we changed the bundling directory to be a directory inside
`outdir` (`cdk.out`). This makes Docker bundling fail during tests.

Fix it by using `fs.realpathSync`.

Closes #10262


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*