chore: migrate constructs to use "constructs" module (#10606)
This commits expands on a previous [commit]. The remaining
CDK modules have been migrated to use the `Construct` and
`IConstruct` from the "constructs" module instead of the
"@aws-cdk/core" module.

Migrating modules to use the "constructs" module, prevents merge
conflicts from arising in these files when the construct compatibility
layer in "@aws-cdk/core" module is removed in the v2 branch.

[commit]: c179699

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*