fix(rds): make ServerlessCluster.secret public
In a last-minute change, the secret was changed from public (matching all of the
other RDS constructs) to private. This means users who don't specify a secret,
and rely on the auto-generated one, have no programmatic means of accessing the
secret.

Brought up by a user on the CDK.dev Slack channel.