fix(rds): secret for ServerlessCluster is not accessible programmatically (#10657)
In a last-minute change, the secret was changed from public (matching all of the
other RDS constructs) to private. This means users who don't specify a secret,
and rely on the auto-generated one, have no programmatic means of accessing the
secret.

Brought up by a user on the CDK.dev Slack channel.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*