fix(ec2): `InitCommand.shellCommand()` renders an argv command instead (#10691)
`InitCommand.shellCommand('abc xyz')` renders to `['abc xyz']` which actually
represents an `argv` command of one element, which is treated
differently. For example, spaces in it are not parsed, which would
be expected for a shell command.

The correct rendering for a shell command is a plain string in the
`command` property.

Fixes #10684.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*