fix(cloudfront-origins): S3Origins with cross-stack buckets cause cyclic references (#10696)
An S3Origin creates an OriginAccessIdentity, and grants that identity read
permissions on the bucket.  This creates cyclic references when the bucket is
located in a different stack than the distribution. The bucket has a dependency
on the generated OAI user for the BucketPolicy, and the Distribution has a
dependency on the bucket's domain name for the origin.

The fix detects a cross-stack bucket and re-parents the OAI in the bucket's
stack to prevent cyclic references.

fixes #10399