chore: migrate all tests off compat layer and use "constructs" module (#10698)
Switching all tests and test utilities to use the "constructs" module
as preparation for v2 branch. As part of CDKv2, the construct compat
layer in the "@aws-cdk/core" module will be removed.

This change pre-migrates tests to the new module so as to reduce code
divergence between the 'v1' and 'v2' code branches. This will in turn
reduce the number of merge conflicts when changes are forward merged
from the 'v1' branch onto the 'v2' branch.

This change also adds a new private tools module - eslint-plugin-cdk -
where CDK specific eslint rules can be added. The first rule now is to
ensure that tests and test related utilities do not use the construct
compat layer.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*