fix(pipelines): cannot use constructs in build environment (#10654)
The environment serialization that is used to calculate a hash
of the project configuration which is used to restart the pipeline
(still following?) was not taking into account that some of the
fields of `Environment` are not just data, but are objects.

We have to turn those into data before we can serialize them.

Fixes #10535.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*