chore(cfnspec): drop step functions patch as the updated spec has restored properties (#10764)
in the `18.5` import, there the AWS::StepFunctions::Activity resource dropped the required
name property and added one called Arn which was unintentional. A patch was added to restore
the resource specification to previous state.

Since #10762 has been merged, those properties have been restored in the resource specification
as of `18.6`. The patch is no longer necessary.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*