chore: remove unnecessary ruby prerequisites
We are not using `ruby` anywhere, and yet we list it in our
prerequisites and check for it in `check-prerequisites.sh`. This makes
it more difficult for new contributors to get started, as this is one
more dependency they must install... for not purpose.

This would enable us to remove `ruby` from the `jsii/superchain` image
in order to make it a little lighter weight (aws/jsii#2092).