fix(lambda): grantInvoke on imported function fails (#10622)
PR #8828 added the ability to add permissions on imported functions, if the
account on the stack and imported function matched. However, when this doesn't
match, attempting to call `grantInvoke` on the imported function results in a
`findChild` error.

This change guards against calling `addPermission` when `canCreatePermissions`
is false by checking if the `CfnPermission` was actually created, and adjusting
the return from `addInvoke` appropriately.

fixes #10607


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*