feat(lambda-python): bundle dependencies in a lambda layer (#9582)
This feature enables `PythonFunction` to include `Pipfile` and `requirements.txt` dependencies in a function and `PythonLayerVersion` to include the dependencies in a lambda layer. The bundling process uses a Dockerfile build to install Pipenv and to cache the dependencies in the Docker layer cache, improving the dev experience by speeding up bundling.

Closes #9406, #9944

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*