feat(config): Scope class for scoping config rules to a specific resource, resource types, tags (#10821)
Previously, we had methods `scopeToResource()`, `scopeToResources()` and `scopeToTag()`
which could be called on a rule. However, only one type of scope restriction could be
added to a rule. Calling these methods sequentially would not result in additive behaviour,
which may be confusing to users.

This change introduces a scope property to config rules with some static convenience methods to
define the scope of a config rule. After this change, only one scope can be defined and behaviour
will be less jarring.

Also added an enum like class `ResourceType` which adds some convenience to selecting resources to scope.

BREAKING CHANGE: `scopeToResource()`, `scopeToResources()` and `scopeToTag()` APIs have been removed. Use the `scope` property to restrict the scope of a Config rule.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*