chore(lambda-event-sources): fix typo in readme (#9959)
>This is a fix for the README in the Kinesis event source example. This removes the undeclared variable `queue` from the example and replaces it with the `stream`.

This is a simple typo pointing to a variable that was not declared in the example for Kinesis.


The motivation for this change is:

- To ensure developers have accurate documentation.
- To ensure developers have a good experience when attempting to understand the features of the CDK from the documentation.

There is a convention in the README's to refer to variables outside the scope of the example e.g.

```ts
myFunction.addEventSource( //etc
```

Where `myFunction` is not declared in the scope of the example. This fix prevents the assumption that a `queue` resource has also been created outside the scope of the example.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*