fix(cli): failure if account cache is malformed (#10887)
Because of concurrent running of integration tests, the account
cache (which is supposed to be a JSON file) can be read in a state
where it's empty or incompletely written, which fails the JSON parse.

If that happens, ignore the error and pretend the cache is empty.

Fixes sporadic concurrency issues in the integration tests.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*