chore(cognito): reorganize identity providers structure so that UserPoolIdentityProviderBase is not exported
we had split out identity providers to have a file per provider.
however, in this organization we were exporting the base class which
we did not intend to.

the `UserPoolIdentityProviderBase` class cannot be moved to a separate
file (i.e. private folder) as JSII will be unable to resolve type.

following our typical pattern of containing all extensions of the base
class within a single file `user-pool-idp.ts` and moved all the tests
over as well.

BREAKING CHANGE: the `UserPoolIdentityProviderBase` abstract class has been removed. Use the `UserPoolIdentityProvider*` classes directly.