chore(cognito): reorganize identity providers structure so that UserPoolIdentityProviderBase is not exported (#10925)
we had split our identity providers to have a file per provider.
however, in this organization we were exporting the base class which
we did not intend to.

the `UserPoolIdentityProviderBase` class cannot be moved to a separate
file (i.e. private folder) as JSII will be unable to resolve type.

updated to follow our typical pattern of containing all extensions of the base
class within a single file `user-pool-idp.ts` and moved all the tests
over as well.

BREAKING CHANGE: the `UserPoolIdentityProviderBase` abstract class has been removed. Use the `UserPoolIdentityProvider*` classes directly.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*