feat(lambda-event-sources): disable source mapping (#10927)
Specifically, it enables this flag in the props of which resources use
an underlying Cfn EventSourceMapping, which carries the Enabled flag
itself.
These are currently: SQS, DynamoDb Streams and Kinesis.

MSK is also listed in the Cfn Docs, but I could not find MSK in `lambda-event-sources`.

Closes #5750

- [X] Added Unit Tests
- [X] Documented addition in `README.md`

This change is not breaking.

I did not add integration tests so far, but could write some if required.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*