fix(core): make Stage compatible to instances from other copies of code
A recent change surfaces an issue where Stage objects from different
code copies don't recognize each other as a Stage. This paht aligns
the way how a Stage determines if something is a Stage to the same
mechanism that App and Stack use.

Thanks to @Shogan and @jogold to narrowing the issue down

fixes #10314