feat(autoscaling): CloudFormation init for ASGs (#9674)
Introduces CloudFormation init (cfn-init) support for autoscaling groups.
This builds on the previous work (#9065) and (#9664) that introduced init
support for instances.

This change also reworks the existing signaling functionality, as this becomes
even more important with cfn-init.

A final change is to export the `CloudFormationInit._attach` method and related
options.

Credit for 90% of this goes to @rix0rrr; all cfn-init support was pair-
programmed, but the ASG stuff was mostly him.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*