refactor(core): simplify cross-stack reference detection
The current design of cross-stack references is needlessly
complicated.

It is based around being able to hook into the Token resolution
mechanism, which is then used to mock-resolve the entire construct tree
and remember all the `CfnReference` objects that were encountered while
doing so.

Then, we poke additional state into each of the encountered tokens so
that *next time* that same token object is resolved in the scope of a
consuming stack, we can output one of the poked values.

This whole solution introduces the following complications:

- Scope-based resolution: instead of resolving to a simple value,
  the token executes code that inspects the scope of the resolver
  to return a different value.
- Time-based resolution: depending on when the token is being
  resolved, it returns something different.

Both of these are the enemy of caching, and of consistency and
predictable behavior.

A better solution to always resolve to a consistent value that
represents the intended special semantics. In that way, the token itself
is trivial (all it needs to do is escape the type system and nothing
else), and template language postprocessors implement the special
features that are necessary.

For references, the issue is that the same token must *sometimes*
evaluate to a `{ Ref }` intrinsic and sometimes must evaluate to a
`{ Fn::ImportValue }` intrinsic.

In this PR, we extend the template language that Tokens render to
from CloudFormation to **CdkFormation**, which is a superset of
CloudFormation that also contains the intrinsic:

```
{ "$Cdk::Ref": ["/path/to/node", "Attr"] }
```

This makes references be able to render to a stable value again,
independent of scope and time.

Finally, in `stack.resolve()` the function `replaceReferences()`
is called which translates from CdkFormation -> CloudFormation,
replacing `{ $Cdk::Ref }` with `{ Ref }` or `{ Fn::ImportValue }`
as the case may be.

It has not yet been fully carried through in this PR, but this change
will be able to get rid of all the fields on `IResolveContext` and all
the resolution hook machinery that make the Token implementation
so complicated.